api/requirements.txt
django
django-rest-framework
django-rest-auth
psycopg2
dj-database-url
gunicorn
django-jet
django-cors-middleware
django-rest-swagger

django-allauth